+++ oauth2_provider/bcp.py
@@ -1,7 +1,7 @@
 """Backward-compatible import shim.
 
 ``oauth2_provider.bcp`` has moved to
-``oauth2_provider.authorization_server.bcp``.
+``oauth2_provider.core.bcp``.
 
 Importing from this old path still works but is deprecated and will be removed
 in django-oauth-toolkit 4.0.
@@ -10,17 +10,15 @@ in django-oauth-toolkit 4.0.
 import sys
 import warnings
 
-from oauth2_provider.authorization_server import bcp as _moved
+from oauth2_provider.core import bcp as _moved
 
 
 warnings.warn(
-    "oauth2_provider.bcp has moved to oauth2_provider.authorization_server.bcp. "
+    "oauth2_provider.bcp has moved to oauth2_provider.core.bcp. "
     "The old import path is deprecated and will be removed in "
     "django-oauth-toolkit 4.0.",
     DeprecationWarning,
     stacklevel=2,
 )
 
-# Alias the old module name to the moved module so every attribute (including
-# private, underscore-prefixed names) resolves with identical object identity.
 sys.modules[__name__] = _moved
